C linked lists print list commit

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -1,5 +1,4 @@
 #include "lists.h"
-#include <stdio.h>
 
 /**
 * print_list- Prints all the elements of a list
@@ -9,7 +8,7 @@
 
 size_t print_list(const list_t *h)
 {
-size_t n = 0;
+unsigned int n = 0;
 
 while (h)
 {
@@ -17,7 +16,7 @@ if (h->str == NULL)
 printf("[0] (nil)\n");
 
 else
-printf("[%d] %s\n", h->len, h->str);
+printf("[%i] %s\n", h->len, h->str);
 
 h = h->next;
 n++;
